PCA plot for RNA-seq data
This is the R code I use to make PCA plots for RNAseq data.
Use this within R studio to create a plot to export.
The input data is a .csv file. The first column of the input data is gene names.
The second column is counts for each gene from an RNA-seq data set.
Each following column is additional data sets you want to compare with their counts for each idicated gene from the first column.
It uses DEseq2 and DEseq2 PCAplotting function edited to be able to plot any PC of your choosing (1,2,3..etc.)